fix: retain model manager singleton across editor
