+++ templates/live2d_parameter_editor.html
@@ -8,6 +8,7 @@
     <title data-i18n="live2d.parameterEditor.title">Live2D 参数编辑器（捏脸系统）</title>
     <script src="/static/i18n-i18next.js"></script>
     <script src="/static/common_dialogs.js"></script>
+    <script src="/static/js/model_manager/named-window-registration.js?v={{ static_asset_version | default('0') }}"></script>
     <link rel="stylesheet" href="/static/css/live2d_parameter_editor.css">
     <link rel="stylesheet" href="/static/css/dark-mode.css">
     <script src="/static/theme-manager.js"></script>
