+++ templates/model_manager.html
@@ -18,6 +18,7 @@
     <!-- i18next 加载器（统一处理 CDN 加载和初始化） -->
     <script src="/static/i18n-i18next.js"></script>
     <script src="/static/common_dialogs.js"></script>
+    <script src="/static/js/model_manager/named-window-registration.js?v={{ static_asset_version | default('0') }}"></script>
     <link rel="stylesheet" href="/static/css/model_manager.css?v={{ static_asset_version | default('0') }}">
     <link rel="stylesheet" href="/static/css/dark-mode.css">
     <link rel="stylesheet" href="/static/libs/driver.min.css?v={{ static_asset_version | default('0') }}">
